feat(publish): refactor checklist and wire preview rail
