+++ src/components/publish/Checklist.tsx
@@ -1,25 +1,90 @@
-import React from 'react';
+import React, { useEffect, useRef, useState } from 'react';
+import { AnimatePresence, motion } from 'framer-motion';
 
-export type ChecklistItem = { id: string; label: string; ok: boolean; required?: boolean };
+export type ChecklistItem = { id: string; label: string; ok: boolean; target?: string };
 
-export default function Checklist({ items }: { items: ChecklistItem[] }) {
-  const allOk = items.every((i) => i.ok || i.required === false);
+type Props =
+  | { items: ChecklistItem[]; onFix?: (target?: string) => void }
+  | { issues: string[]; onFix?: (target?: string) => void };
+
+export default function Checklist(props: Props) {
+  const items: ChecklistItem[] = 'items' in props
+    ? props.items
+    : props.issues.map((label, i) => ({ id: String(i), label, ok: false }));
+  const onFix = 'onFix' in props ? props.onFix : undefined;
+
+  const prev = useRef<Record<string, boolean>>({});
+  const [announcement, setAnnouncement] = useState('');
+
+  useEffect(() => {
+    const turnedGreen = items.filter((i) => i.ok && prev.current[i.id] === false);
+    if (turnedGreen.length > 0) {
+      setAnnouncement(`${turnedGreen[0].label} resolved`);
+    }
+    items.forEach((i) => {
+      prev.current[i.id] = i.ok;
+    });
+  }, [items]);
+
+  const handleFix = (target?: string) => {
+    if (onFix) {
+      onFix(target);
+      return;
+    }
+    if (!target) return;
+    const el = document.getElementById(target) as HTMLElement | null;
+    el?.scrollIntoView({ behavior: 'smooth', block: 'center' });
+    (el?.querySelector('input,select,textarea,button,[tabindex]') as HTMLElement | null)?.focus?.();
+  };
+
+  const allGood = items.every((i) => i.ok);
+
   return (
-    <div className="mb-4" role="region" aria-label="Compliance checklist">
-      <div className="flex items-center justify-between">
-        <h2 className="text-sm font-semibold text-slate-800">Publication‑ready checklist</h2>
-        <span className={`text-xs px-2 py-0.5 rounded ${allOk ? 'bg-green-100 text-green-800' : 'bg-amber-100 text-amber-800'}`}>
-          {allOk ? 'All good' : 'Finish required items'}
-        </span>
-      </div>
-      <ul className="mt-2 grid grid-cols-1 md:grid-cols-2 gap-x-6 gap-y-1 text-sm">
-        {items.map((it) => (
-          <li key={it.id} className="flex items-start gap-2">
-            <span aria-hidden className={`mt-0.5 inline-block w-2.5 h-2.5 rounded-full ${it.ok ? 'bg-green-600' : 'bg-amber-600'}`} />
-            <span className={it.ok ? 'text-slate-700' : 'text-amber-800'}>{it.label}</span>
-          </li>
-        ))}
+    <div className="rounded-xl border bg-white p-5 shadow-sm" role="region" aria-label="Compliance checklist">
+      <h3 className="mb-3 text-sm font-semibold tracking-tight">Compliance checklist</h3>
+      <ul className="space-y-2 text-sm">
+        <AnimatePresence initial={false}>
+          {allGood && (
+            <motion.li
+              layout
+              key="all-good"
+              className="flex items-center gap-2 text-emerald-700"
+            >
+              <span aria-hidden className="inline-flex h-4 w-4 items-center justify-center rounded-full bg-emerald-600 text-white text-[10px]">✔</span>
+              All checks passed — ready to publish
+            </motion.li>
+          )}
+          {!allGood &&
+            items.map((item) => (
+              <motion.li
+                layout
+                key={item.id}
+                className="flex items-center justify-between"
+              >
+                <div className="flex items-center gap-2">
+                  <span
+                    aria-hidden
+                    className={`inline-flex h-4 w-4 items-center justify-center rounded-full text-[10px] ${item.ok ? 'bg-emerald-600 text-white' : 'bg-amber-500 text-white'}`}
+                  >
+                    {item.ok ? '✔' : '✘'}
+                  </span>
+                  <span className="text-slate-800">{item.label}</span>
+                </div>
+                {!item.ok && (
+                  <button
+                    className="ml-4 rounded-md border px-2 py-1 text-xs focus-visible:outline-none focus-visible:ring-2 focus-visible:ring-offset-2 focus-visible:ring-blue-500"
+                    onClick={() => handleFix(item.target)}
+                  >
+                    Fix this
+                  </button>
+                )}
+              </motion.li>
+            ))}
+        </AnimatePresence>
       </ul>
+      <div className="sr-only" aria-live="polite">
+        {announcement}
+      </div>
     </div>
   );
 }
